pyproject.toml
[project]
name = "bardagent"
version = "0.1.0"
description = "Assessment Agent for Bardeen.ai Recruitment Process"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "langgraph>=0.2.35",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "playwright>=1.57.0",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "wikipedia>=1.4.0",
    "yfinance>=1.0",
    "youtube-search>=2.2.0",
    "tqdm>=4.67.1",
    "sympy>=1.14.0",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.7",
    "tenacity>=9.1.2",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "pillow>=12.1.0",
    "youtube-transcript-api>=1.2.4",
    "google-generativeai>=0.8.6",
    "gradio>=6.5.1",
    "google-genai>=1.61.0",
]
